# Conflicts:
#	MANIFEST.in
#	PPOCRLabel/PPOCRLabel.py
#	PPOCRLabel/README.md
#	PPOCRLabel/README_ch.md
#	PPOCRLabel/libs/autoDialog.py
#	PPOCRLabel/requirements.txt
#	PPOCRLabel/setup.py
#	README.md
#	README_ch.md
#	applications/README.md
#	deploy/paddlejs/README.md
#	deploy/paddlejs/README_ch.md
#	deploy/slim/quantization/README.md
#	deploy/slim/quantization/export_model.py
#	doc/doc_ch/PP-OCRv3_det_train.md
#	doc/doc_ch/algorithm_det_east.md
#	doc/doc_ch/finetune.md
#	doc/doc_ch/quickstart.md
#	doc/doc_en/finetune_en.md
#	doc/doc_en/models_list_en.md
#	doc/doc_en/ocr_book_en.md
#	doc/doc_en/quickstart_en.md
#	paddleocr.py
#	requirements.txt
#	setup.py
#	train.sh
#	unstructured_paddleocr/paddle_tools/infer/predict_system.py
#	unstructured_paddleocr/paddle_tools/infer/utility.py
#	unstructured_paddleocr/ppocr/data/imaug/__init__.py
#	unstructured_paddleocr/ppocr/data/imaug/abinet_aug.py
#	unstructured_paddleocr/ppocr/data/imaug/label_ops.py
#	unstructured_paddleocr/ppocr/data/imaug/rec_img_aug.py
#	unstructured_paddleocr/ppocr/data/multi_scale_sampler.py
#	unstructured_paddleocr/ppocr/losses/rec_aster_loss.py
#	unstructured_paddleocr/ppocr/losses/rec_cppd_loss.py
#	unstructured_paddleocr/ppocr/losses/rec_nrtr_loss.py
#	unstructured_paddleocr/ppocr/losses/rec_parseq_loss.py
#	unstructured_paddleocr/ppocr/losses/rec_satrn_loss.py
#	unstructured_paddleocr/ppocr/modeling/backbones/det_pp_lcnet_v2.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_hgnet.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_lcnetv3.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_repvit.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_shallow_cnn.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_svtrnet.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_svtrv2.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_vit.py
#	unstructured_paddleocr/ppocr/modeling/backbones/rec_vit_parseq.py
#	unstructured_paddleocr/ppocr/modeling/heads/rec_cppd_head.py
#	unstructured_paddleocr/ppocr/modeling/heads/rec_nrtr_head.py
#	unstructured_paddleocr/ppocr/modeling/heads/rec_parseq_head.py
#	unstructured_paddleocr/ppocr/modeling/heads/rec_satrn_head.py
#	unstructured_paddleocr/ppocr/modeling/necks/intracl.py
#	unstructured_paddleocr/ppocr/postprocess/db_postprocess.py
#	unstructured_paddleocr/ppocr/utils/dict/README.md
#	unstructured_paddleocr/ppocr/utils/dict/bengali_dict.txt
#	unstructured_paddleocr/ppocr/utils/dict/bm_dict.txt
#	unstructured_paddleocr/ppocr/utils/dict/gujarati_dict.txt
#	unstructured_paddleocr/ppocr/utils/dict/kazakh_dict.txt
#	unstructured_paddleocr/ppocr/utils/dict/parseq_dict.txt
#	unstructured_paddleocr/ppocr/utils/loggers/vdl_logger.py
#	unstructured_paddleocr/ppocr/utils/utility.py
#	unstructured_paddleocr/ppstructure/kie/README.md
#	unstructured_paddleocr/ppstructure/pdf2word/pdf2word.py
#	unstructured_paddleocr/ppstructure/predict_system.py
#	unstructured_paddleocr/ppstructure/recovery/recovery_to_doc.py
#	unstructured_paddleocr/ppstructure/return_word_pos.md
#	unstructured_paddleocr/ppstructure/utility.py

+++ deploy/avh/convert_image.py
@@ -24,13 +24,13 @@ import math
 from PIL import Image
 import numpy as np
 
+
 def resize_norm_img(img, image_shape, padding=True):
     imgC, imgH, imgW = image_shape
     h = img.shape[0]
     w = img.shape[1]
     if not padding:
-        resized_image = cv2.resize(
-            img, (imgW, imgH), interpolation=cv2.INTER_LINEAR)
+        resized_image = cv2.resize(img, (imgW, imgH), interpolation=cv2.INTER_LINEAR)
         resized_w = imgW
     else:
         ratio = w / float(h)
@@ -39,7 +39,7 @@ def resize_norm_img(img, image_shape, padding=True):
         else:
             resized_w = int(math.ceil(imgH * ratio))
         resized_image = cv2.resize(img, (resized_w, imgH))
-    resized_image = resized_image.astype('float32')
+    resized_image = resized_image.astype("float32")
     if image_shape[0] == 1:
         resized_image = resized_image / 255
         resized_image = resized_image[np.newaxis, :]
@@ -80,16 +80,16 @@ def create_headers(image_name):
 
     # Resize image to 32x320
     img = cv2.imread(img_path)
-    img = resize_norm_img(img, [3,32,320])
+    img = resize_norm_img(img, [3, 32, 320])
     img_data = img.astype("float32")
-    
+
     # # Add the batch dimension, as we are expecting 4-dimensional input: NCHW.
     img_data = np.expand_dims(img_data, axis=0)
 
     # Create input header file
     create_header_file("inputs", "input", img_data, "./include")
     # Create output header file
-    output_data = np.zeros([7760], np.float)
+    output_data = np.zeros([7760], np.float32)
     create_header_file(
         "outputs",
         "output",
